+++ src/connection_settings.h
@@ -16,20 +16,15 @@ struct ConnectionSettings {
   int audio_topic_ch = 1;
 #endif
 
-  bool no_video = false;
-  bool no_audio = false;
+  bool no_video_device = false;
+  bool no_audio_device = false;
   bool force_i420 = false;
   bool use_native = false;
   std::string video_device = "";
-  std::string video_codec = "VP8";
-  std::string audio_codec = "OPUS";
-  int video_bitrate = 0;
-  int audio_bitrate = 0;
   std::string resolution = "VGA";
   int framerate = 30;
   bool fixed_resolution = false;
   std::string priority = "BALANCE";
-  int port = 8080;
   bool use_sdl = false;
   bool show_me = false;
   int window_width = 640;
@@ -41,14 +36,24 @@ struct ConnectionSettings {
 
   std::string sora_signaling_host = "wss://example.com/signaling";
   std::string sora_channel_id;
+  bool sora_video = true;
+  bool sora_audio = true;
+  // 空文字の場合コーデックは Sora 側で決める
+  std::string sora_video_codec = "";
+  std::string sora_audio_codec = "";
+  // 0 の場合ビットレートは Sora 側で決める
+  int sora_video_bitrate = 0;
+  int sora_audio_bitrate = 0;
   bool sora_auto_connect = false;
   nlohmann::json sora_metadata;
   // upstream or downstream
   std::string sora_role = "upstream";
   bool sora_multistream = false;
   int sora_spotlight = -1;
+  int sora_port = -1;
 
   std::string test_document_root;
+  int test_port = 8080;
 
   std::string ayame_signaling_host;
   std::string ayame_room_id;
@@ -101,27 +106,30 @@ struct ConnectionSettings {
 
   friend std::ostream& operator<<(std::ostream& os,
                                   const ConnectionSettings& cs) {
-    os << "no_video: " << (cs.no_video ? "true" : "false") << "\n";
-    os << "no_audio: " << (cs.no_audio ? "true" : "false") << "\n";
-    os << "video_codec: " << cs.video_codec << "\n";
-    os << "audio_codec: " << cs.audio_codec << "\n";
-    os << "video_bitrate: " << cs.video_bitrate << "\n";
-    os << "audio_bitrate: " << cs.audio_bitrate << "\n";
+    os << "no_video_device: " << (cs.no_video_device ? "true" : "false")
+       << "\n";
+    os << "no_audio_device: " << (cs.no_audio_device ? "true" : "false")
+       << "\n";
     os << "resolution: " << cs.resolution << "\n";
     os << "framerate: " << cs.framerate << "\n";
     os << "fixed_resolution: " << (cs.fixed_resolution ? "true" : "false")
        << "\n";
     os << "priority: " << cs.priority << "\n";
-    os << "port: " << cs.port << "\n";
     os << "ayame_signaling_host: " << cs.ayame_signaling_host << "\n";
     os << "ayame_room_id: " << cs.ayame_room_id << "\n";
     os << "ayame_client_id: " << cs.ayame_client_id << "\n";
     os << "sora_signaling_host: " << cs.sora_signaling_host << "\n";
     os << "sora_channel_id: " << cs.sora_channel_id << "\n";
+    os << "sora_video_codec: " << cs.sora_video_codec << "\n";
+    os << "sora_audio_codec: " << cs.sora_audio_codec << "\n";
+    os << "sora_video_bitrate: " << cs.sora_video_bitrate << "\n";
+    os << "sora_audio_bitrate: " << cs.sora_audio_bitrate << "\n";
     os << "sora_auto_connect: " << (cs.sora_auto_connect ? "true" : "false")
        << "\n";
     os << "sora_metadata: " << cs.sora_metadata << "\n";
+    os << "sora_port: " << cs.sora_port << "\n";
     os << "test_document_root: " << cs.test_document_root << "\n";
+    os << "test_port: " << cs.test_port << "\n";
     return os;
   }
 };